fixed .travis.yml

+++ .travis.sh
@@ -7,7 +7,7 @@ github_name="Travis CI"
 github_mail="travis@travis-ci.org"
 branch="gh-pages"
 deploy_dir=$(mktemp -d)
-repo_url="https://${GITHUB_TOKEN"}@github.com/${TRAVIS_REPO_SLUG}.git"
+repo_url="https://${GITHUB_TOKEN}@github.com/${TRAVIS_REPO_SLUG}.git"
 
 
 git config --global user.name "${github_name}"
@@ -17,7 +17,7 @@ git clone --depth 1 --branch "${branch}" "${repo_url}" "${deploy_dir}"
 
 
 
-rm -rf "${deploy_dir}/${TRAVIS_BRANCH}"
+rm -rf "${deploy_dir:?}/${TRAVIS_BRANCH}"
 cp -rv dist "${deploy_dir}/${TRAVIS_BRANCH}"
 
 pushd "${deploy_dir}"
